move DAQ to device
